refactor: feature tests
