+++ features/useCases/MoveNext/moveNext.feature
@@ -1,0 +1,6 @@
+@moveNext
+Feature: Atualizar pedido na fila
+
+    @atualizarPedido
+    Scenario: Atualiza posição do pedido passando o ID
+
